pyproject.toml
# https://github.com/python-poetry/poetry#introduction

[tool.poetry]
name = "weatherbot"
version = "0.0.0"
description = ""
authors = []


[tool.poetry.dependencies]
python = "~3.8"

botx = { git = "https://github.com/ExpressApp/pybotx.git", branch = "feature/smartapp" }

fastapi = "~0.61.2"
uvicorn = "~0.12.3"
python-multipart = "~0.0.5"

tortoise-orm = "~0.16.18"
databases = { version = "~0.4.1", extras = ["postgresql"] }
alembic = "~1.3.2"
psycopg2-binary = "~2.8.6"  # for alembic

loguru = "~0.5.3"
mako = "~1.1.0"

sentry-sdk = "~0.17.3"
aiofiles = "^0.6.0"


[tool.poetry.dev-dependencies]
black = "~20.8b1"
isort = "4.3.21"
autoflake = "~1.4"
mypy = "~0.782"
wemake-python-styleguide = "~0.14"

docker = "~4.0"
asgi-lifespan = "~1.0.0"

pytest = "~5.4.1"
pytest-cov = "~2.8.1"
pytest-asyncio = "~0.12.0"
pytest-localserver = "~0.5.0"
python-dotenv = "~0.13.0"
pytest-env = "~0.6.2"
pytest-clarity = "~0.3.0-alpha.0"
pytest-timeout = "~1.3.4"

molten = "1.0.1"
redis = "3.5.3"
tenacity = "6.2.0"


[build-system]
requires = ["poetry>=1.0"]
build-backend = "poetry.masonry.api"
